final app v1.0.0

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,20 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -40,12 +50,30 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+        // Handle file upload 
+         if($request->hasFile('cover_image')){
+            // get file name with the extention 
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just file name 
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // get just extention 
+            $extention = $request->file('cover_image')->getClientOriginalExtension();
+            // Create File Name to store 
+            $FileNameToStore = $filename.'_'.time().'.'.$extention;
+            // upload the image 
+            $path =  $request->file('cover_image')->storeAs('Public/cover_images',$FileNameToStore);
+         } else {
+             $FileNameToStore = 'noimage.jpg'; 
+         }
         // create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $FileNameToStore;
         $post->save();
         return redirect('/posts')->with('success', 'Post Created');
     }
@@ -70,7 +98,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        //check if current user 
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id ){
+            return redirect('/posts')->with('error','Unauthorised Page');   
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -85,12 +117,33 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload 
+        if($request->hasFile('cover_image')){
+            // get file name with the extention 
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just file name 
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // get just extention 
+            $extention = $request->file('cover_image')->getClientOriginalExtension();
+            // Create File Name to store 
+            $FileNameToStore = $filename.'_'.time().'.'.$extention;
+            // upload the image 
+            $path =  $request->file('cover_image')->storeAs('Public/cover_images',$FileNameToStore);
+         } else {
+             $FileNameToStore = 'noimage.jpg'; 
+         }
+
         // create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+        $post->cover_image = $FileNameToStore;
+        }
         $post->save();
         return redirect('/posts')->with('success', 'Post Updated');
     }
@@ -104,6 +157,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id ){
+            return redirect('/posts')->with('error','Unauthorised Page');   
+        }
+        if($post->cover_image != 'noimage.jpg'){
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
